Fix all broken specs

+++ app/controllers/videos.js
@@ -5,34 +5,40 @@
  * Created by henryehly on 2017/01/18.
  */
 
-const services        = require('../services');
-const ResponseWrapper = services.ResponseWrapper;
-const db              = require('../models');
-const ModelHelper     = services.Model(db);
-const Subcategory     = db.Subcategory;
-const CuedVideo       = db.CuedVideo;
-const Speaker         = db.Speaker;
-const Video           = db.Video;
-const Like            = db.Like;
 const k               = require('../../config/keys.json');
+const services        = require('../services');
+const ResponseWrapper = services['ResponseWrapper'];
+const GetNativeError  = services['GetNativeError'];
+const db              = require('../models');
+const ModelHelper     = services['Model'](db);
+const Subcategory     = db[k.Model.Subcategory];
+const CuedVideo       = db[k.Model.CuedVideo];
+const Language        = db[k.Model.Language];
+const Speaker         = db[k.Model.Speaker];
+const Video           = db[k.Model.Video];
+const Like            = db[k.Model.Like];
 
 const Promise         = require('bluebird');
 
 module.exports.index = (req, res, next) => {
-    const conditions = {language_code: req.query.lang || 'en'};
+    const conditions = {};
 
-    return Subcategory.findIdsForCategoryIdOrSubcategoryId(req.query).then(subcategoryIds => {
+    return Language.findOne({where: {code: req.body.lang || 'en'}}).then(language => {
+        conditions.language_id = language.get(k.Attr.Id);
+
+        return Subcategory.findIdsForCategoryIdOrSubcategoryId(req.query);
+    }).then(subcategoryIds => {
         if (subcategoryIds.length) {
             conditions.subcategory_id = {$in: subcategoryIds};
         }
 
         const createdAt  = ModelHelper.getDateAttrForTableColumnTZOffset(k.Model.Video, k.Attr.CreatedAt, req.query.time_zone_offset);
-        const cued       = Video.getCuedAttributeForUserId(req.userId);
+        const cued       = Video.getCuedAttributeForUserId(req.user[k.Attr.Id]);
         const attributes = [createdAt, k.Attr.Id, k.Attr.LoopCount, k.Attr.PictureUrl, k.Attr.VideoUrl, k.Attr.Length, cued];
 
         const scopes = [
             'newestFirst',
-            {method: ['cuedAndMaxId', req.query.cued_only, req.userId, req.query.max_id]},
+            {method: ['cuedAndMaxId', req.query.cued_only, req.user[k.Attr.Id], req.query.max_id]},
             {method: ['count', req.query.count]},
             {method: ['includeSubcategoryNameAndId', Subcategory]},
             {method: ['includeSpeakerName', Speaker]}
@@ -55,11 +61,11 @@ module.exports.index = (req, res, next) => {
 
 module.exports.show = (req, res, next) => {
     const likeCount = Video.getLikeCount(db, req.params.id);
-    const liked     = Video.isLikedByUser(db, req.params.id, req.userId);
-    const cued      = Video.isCuedByUser(db, req.params.id, req.userId);
+    const liked     = Video.isLikedByUser(db, req.params.id, req.user[k.Attr.Id]);
+    const cued      = Video.isCuedByUser(db, req.params.id, req.user[k.Attr.Id]);
 
     const relatedCreatedAt = ModelHelper.getDateAttrForTableColumnTZOffset(k.Model.Video, k.Attr.CreatedAt, req.query.time_zone_offset);
-    const relatedCued      = Video.getCuedAttributeForUserId(req.userId);
+    const relatedCued      = Video.getCuedAttributeForUserId(req.user[k.Attr.Id]);
 
     const relatedVideos = Video.scope([
         {method: ['includeSubcategoryNameAndId', Subcategory]},
@@ -68,15 +74,17 @@ module.exports.show = (req, res, next) => {
         'orderByRandom'
     ]).findAll({
         attributes: [k.Attr.Id, relatedCreatedAt, k.Attr.Length, k.Attr.PictureUrl, k.Attr.LoopCount, relatedCued],
+        include: [{model: Language, attributes: [k.Attr.Name, k.Attr.Code], as: 'language'}],
         limit: 3
     }).catch(next);
 
-    const video = Video.scope('includeTranscripts').findById(+req.params.id, {
+    const video = Video.scope('includeTranscripts').findById(+req.params[k.Attr.Id], {
         include: [
             {model: Speaker, attributes: [k.Attr.Id, k.Attr.Description, k.Attr.Name, k.Attr.PictureUrl], as: 'speaker'},
-            {model: Subcategory, attributes: [k.Attr.Id, k.Attr.Name], as: 'subcategory'}
+            {model: Subcategory, attributes: [k.Attr.Id, k.Attr.Name], as: 'subcategory'},
+            {model: Language, attributes: [k.Attr.Name, k.Attr.Code], as: 'language'}
         ],
-        attributes: [k.Attr.Description, k.Attr.Id, k.Attr.LoopCount, k.Attr.PictureUrl, k.Attr.VideoUrl, k.Attr.Length, k.Attr.LanguageCode]
+        attributes: [k.Attr.Description, k.Attr.Id, k.Attr.LoopCount, k.Attr.PictureUrl, k.Attr.VideoUrl, k.Attr.Length]
     }).catch(next);
 
     return Promise.join(likeCount, liked, cued, relatedVideos, video, (likeCount, liked, cued, relatedVideos, video) => {
@@ -102,10 +110,14 @@ module.exports.show = (req, res, next) => {
 };
 
 module.exports.like = (req, res, next) => {
-    return Video.findById(+req.params.id).then(video => {
+    return Video.findById(parseInt(req.params[k.Attr.Id])).then(video => {
+        if (!video) {
+            throw new GetNativeError(k.Error.ResourceNotFound);
+        }
+
         return Like.create({
-            video_id: video.id,
-            user_id: req.userId
+            video_id: video[k.Attr.Id],
+            user_id: req.user[k.Attr.Id]
         });
     }).then(() => {
         res.sendStatus(204);
@@ -116,11 +128,15 @@ module.exports.like = (req, res, next) => {
 };
 
 module.exports.unlike = (req, res, next) => {
-    return Video.findById(+req.params.id).then(video => {
+    return Video.findById(parseInt(req.params[k.Attr.Id])).then(video => {
+        if (!video) {
+            throw new GetNativeError(k.Error.ResourceNotFound);
+        }
+
         return Like.destroy({
             where: {
-                video_id: video.id,
-                user_id: req.userId
+                video_id: video[k.Attr.Id],
+                user_id: req.user[k.Attr.Id]
             },
             limit: 1
         });
@@ -135,7 +151,7 @@ module.exports.unlike = (req, res, next) => {
 module.exports.queue = (req, res, next) => {
     return CuedVideo.create({
         video_id: req.params.id,
-        user_id: req.userId
+        user_id: req.user[k.Attr.Id]
     }).then(() => {
         res.sendStatus(204);
     }).catch(next);
@@ -145,7 +161,7 @@ module.exports.dequeue = (req, res, next) => {
     return CuedVideo.destroy({
         where: {
             video_id: req.params.id,
-            user_id: req.userId
+            user_id: req.user[k.Attr.Id]
         },
         limit: 1
     }).then(() => {
